// src/service/enums/service-enums.ts
export enum ServiceEnum {
  SERVICE_EXIST = 'Dịch vụ đã tồn tại.',
  SERVICE_CREATE_ERROR = 'Không thể tạo dịch vụ.',
  SERVICE_NOT_EXIST = 'Dịch vụ không tồn tại.',
  AUTHORIZED_ADMIN = 'Chỉ admin mới có thể sử dụng tính năng này.',
  STATUS_FORMAT = 'Định dạng không đúng.',
}

export enum TicketTypeEnum {
  REQUEST_NEW_SERVICE = 'Yêu cầu dịch vụ mới',
  REQUEST_FIX = 'Yêu cầu sửa chữa',
  REQUEST_OTHER = 'Khác',
}